+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.6] - 2023-01-11
+
+- for Scalar tests
+  - the comparison respects the type of the expression and result (string or number)
+  - result can be an expression as well
+- the json schema can be generated directly from the typescript interfaces/types
+- re-write of the EventBus to correctly define the emitted types
+
 ## [0.0.6] - 2023-01-10
 
 - Initial release